final version

+++ YemekhaneApp.Application/CQRS/Commands/MealRecord/DeleteMealRecordCommand.cs
@@ -6,11 +6,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using YemekhaneApp.Application.Interfaces;
-using MealRecordEntity= YemekhaneApp.Domain.Entities.MealRecord;
+using MealRecordEntity = YemekhaneApp.Domain.Entities.MealRecord;
 
 namespace YemekhaneApp.Application.CQRS.Commands.MealRecord
 {
-    public class DeleteMealRecordCommand:IRequest<ServiceResponse<Guid>>
+    public class DeleteMealRecordCommand : IRequest<ServiceResponse<Guid>>
     {
         public Guid Id { get; set; }
         public DeleteMealRecordCommand(Guid id)
@@ -26,18 +26,35 @@ namespace YemekhaneApp.Application.CQRS.Commands.MealRecord
             }
             public async Task<ServiceResponse<Guid>> Handle(DeleteMealRecordCommand request, CancellationToken cancellationToken)
             {
+                await using var transaction = await _unitOfWork.BeginTransactionAsync();
                 try
                 {
                     var mealRecordRepo = _unitOfWork.GetRepository<MealRecordEntity>();
                     var mealRecord = await mealRecordRepo.GetByGuidAsync(request.Id);
                     if (mealRecord == null)
                         return new ServiceResponse<Guid>("Meal record not found.");
+
+                    // İlişkili kullanıcıyı getir
+                    var employeeRepo = _unitOfWork.GetRepository<YemekhaneApp.Domain.Entities.Employee>();
+                    var employee = await employeeRepo.GetByGuidAsync(mealRecord.EmployeeId);
+                    if (employee == null)
+                        return new ServiceResponse<Guid>("İlişkili çalışan bulunamadı.");
+
+                    // mealCount -1
+                    if (employee.TotalMealCount > 0)
+                        employee.TotalMealCount--;
+
+                    await employeeRepo.UpdateAsync(employee);
                     await mealRecordRepo.DeleteAsync(mealRecord);
+
                     await _unitOfWork.SaveAsync();
+                    await transaction.CommitAsync();
+
                     return new ServiceResponse<Guid>(mealRecord.Id);
                 }
                 catch (Exception ex)
                 {
+                    await transaction.RollbackAsync();
                     return new ServiceResponse<Guid>($"Error deleting meal record: {ex.Message}");
                 }
             }
